+++ tests/test_widerface_format.py
@@ -16,36 +16,37 @@ class WiderFaceFormatTest(TestCase):
                 annotations=[
                     Bbox(0, 2, 4, 2),
                     Bbox(0, 1, 2, 3, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 2, 'invalid': 0}),
+                        'blur': '2', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '2', 'invalid': '0'}),
                     Label(0),
                 ]
             ),
             DatasetItem(id='2', subset='train', image=np.ones((10, 10, 3)),
                 annotations=[
                     Bbox(0, 2, 4, 2, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 1,
-                        'occluded': 0, 'pose': 1, 'invalid': 0}),
+                        'blur': '2', 'expression': '0', 'illumination': '1',
+                        'occluded': '0', 'pose': '1', 'invalid': '0'}),
                     Bbox(3, 3, 2, 3, attributes = {
-                        'blur': 0, 'expression': 1, 'illumination': 0,
-                        'occluded': 0, 'pose': 2, 'invalid': 0}),
+                        'blur': '0', 'expression': '1', 'illumination': '0',
+                        'occluded': '0', 'pose': '2', 'invalid': '0'}),
                     Bbox(2, 1, 2, 3, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 0, 'invalid': 1}),
+                        'blur': '2', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '0', 'invalid': '1'}),
                     Label(1),
                 ]
             ),
 
             DatasetItem(id='3', subset='val', image=np.ones((8, 8, 3)),
                 annotations=[
-                    Bbox(0, 1, 5, 2, attributes = {
-                        'blur': 2, 'expression': 1, 'illumination': 0,
-                        'occluded': 0, 'pose': 1, 'invalid': 0}),
-                    Bbox(0, 2, 3, 2),
+                    Bbox(0, 1.1, 5.3, 2.1, attributes = {
+                        'blur': '2', 'expression': '1', 'illumination': '0',
+                        'occluded': '0', 'pose': '1', 'invalid': '0'}),
+                    Bbox(0, 2, 3, 2, attributes = {
+                        'occluded': 'False'}),
                     Bbox(0, 2, 4, 2),
                     Bbox(0, 7, 3, 2, attributes = {
-                        'blur': 2, 'expression': 1, 'illumination': 0,
-                        'occluded': 0, 'pose': 1, 'invalid': 0}),
+                        'blur': '2', 'expression': '1', 'illumination': '0',
+                        'occluded': '0', 'pose': '1', 'invalid': '0'}),
                 ]
             ),
 
@@ -65,13 +66,16 @@ class WiderFaceFormatTest(TestCase):
         source_dataset = Dataset.from_iterable([
             DatasetItem(id='a/b/1', image=np.ones((8, 8, 3)),
                 annotations=[
-                    Bbox(0, 2, 4, 2),
-                    Bbox(0, 1, 2, 3, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 2, 'invalid': 0}),
+                    Bbox(0, 2, 4, 2, label=2),
+                    Bbox(0, 1, 2, 3, label=1, attributes = {
+                        'blur': '2', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '2', 'invalid': '0'}),
                 ]
             ),
-        ], categories=[])
+        ], categories={
+            AnnotationType.label: LabelCategories.from_iterable(
+                'label_' + str(i) for i in range(3)),
+        })
 
         with TestDir() as test_dir:
             WiderFaceConverter.convert(source_dataset, test_dir, save_images=True)
@@ -85,10 +89,10 @@ class WiderFaceFormatTest(TestCase):
                 annotations=[
                     Bbox(0, 2, 4, 2),
                     Bbox(0, 1, 2, 3, attributes = {
-                        'non-widerface attribute': 0,
-                        'blur': 1, 'invalid': 1}),
+                        'non-widerface attribute': '0',
+                        'blur': 1, 'invalid': '1'}),
                     Bbox(1, 1, 2, 2, attributes = {
-                        'non-widerface attribute': 0}),
+                        'non-widerface attribute': '0'}),
                 ]
             ),
         ], categories=[])
@@ -98,7 +102,7 @@ class WiderFaceFormatTest(TestCase):
                 annotations=[
                     Bbox(0, 2, 4, 2),
                     Bbox(0, 1, 2, 3, attributes = {
-                        'blur': 1, 'invalid': 1}),
+                        'blur': '1', 'invalid': '1'}),
                     Bbox(1, 1, 2, 2),
                 ]
             ),
@@ -122,36 +126,36 @@ class WiderFaceImporterTest(TestCase):
                 image=np.ones((10, 15, 3)),
                 annotations=[
                     Bbox(1, 2, 2, 2, attributes = {
-                        'blur': 0, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 0, 'invalid': 0}),
-                        Label(0),
+                        'blur': '0', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '0', 'invalid': '0'}),
+                    Label(0),
                 ]
             ),
             DatasetItem(id='1_Handshaking_image_02', subset='train',
                 image=np.ones((10, 15, 3)),
                 annotations=[
                     Bbox(1, 1, 2, 2, attributes = {
-                        'blur': 0, 'expression': 0, 'illumination': 1,
-                        'occluded': 0, 'pose': 0, 'invalid': 0}),
+                        'blur': '0', 'expression': '0', 'illumination': '1',
+                        'occluded': '0', 'pose': '0', 'invalid': '0'}),
                     Bbox(5, 1, 2, 2, attributes = {
-                        'blur': 0, 'expression': 0, 'illumination': 1,
-                        'occluded': 0, 'pose': 0, 'invalid': 0}),
-                        Label(1),
+                        'blur': '0', 'expression': '0', 'illumination': '1',
+                        'occluded': '0', 'pose': '0', 'invalid': '0'}),
+                    Label(1),
                 ]
             ),
             DatasetItem(id='0_Parade_image_03', subset='val',
                 image=np.ones((10, 15, 3)),
                 annotations=[
                     Bbox(0, 0, 1, 1, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 2, 'invalid': 0}),
+                        'blur': '2', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '2', 'invalid': '0'}),
                     Bbox(3, 2, 1, 2, attributes = {
-                        'blur': 0, 'expression': 0, 'illumination': 0,
-                        'occluded': 1, 'pose': 0, 'invalid': 0}),
+                        'blur': '0', 'expression': '0', 'illumination': '0',
+                        'occluded': '1', 'pose': '0', 'invalid': '0'}),
                     Bbox(5, 6, 1, 1, attributes = {
-                        'blur': 2, 'expression': 0, 'illumination': 0,
-                        'occluded': 0, 'pose': 2, 'invalid': 0}),
-                        Label(0),
+                        'blur': '2', 'expression': '0', 'illumination': '0',
+                        'occluded': '0', 'pose': '2', 'invalid': '0'}),
+                    Label(0),
                 ]
             ),
         ], categories= ['Parade', 'Handshaking'])